backend/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.24.1
pyjwt==2.8.0
python-multipart==0.0.6
bcrypt==4.1.1
passlib==1.7.4
pymongo==4.6.0
motor==3.3.2
aiofiles==23.2.1
email-validator==2.1.1
supabase==2.1.0
redis==5.0.1
websockets==12.0
python-jose==3.3.0
pillow==10.0.0
aiosmtplib==3.0.0
jinja2==3.1.2
cryptography==41.0.7
openai==1.3.0
